moved excel building to documentation only
